Fix #28: BufferCache::compact too expensive
Remove the vector::erase and only clear the buffer allocation. Requires
keeping track of allocated commands. Refactor code to use atomic ssize_t
for size variables (was a_int32_t).

+++ co/bufferCache.h
@@ -39,7 +39,7 @@ namespace detail { class BufferCache; }
     class BufferCache
     {
     public:
-        CO_API BufferCache( const int32_t minFree );
+        CO_API BufferCache( const ssize_t minFree );
         CO_API ~BufferCache();
 
         /** @return a new buffer. */
